feat: kitty keyboard protocol (levels 1-5) + password input detection
- Widen mode_flags from u8 to u16 for kitty keyboard levels
- Add MODE_KITTY_REPORT_EVENTS/ALTERNATES/ALL/TEXT constants
- Encode keys with CSI u + kitty extensions per active level
- Detect password input via PTY termios (ICANON+!ECHO)
- Suppress broadcast mode and prediction during password input
- Bump wire protocol version to 2

+++ crates/ciri-app/src/prediction.rs
@@ -88,7 +88,7 @@ pub struct GridInfo {
     pub cursor_col: u16,
     pub cols: u16,
     pub rows: u16,
-    pub mode_flags: u8,
+    pub mode_flags: u16,
 }
 
 impl PredictionEngine {
